Rewrite TistoryError, the TistoryResponse wrapper and TistoryRequest
TistoryError now has attributes for status_code, error_message and error_type
_error_handler() was moved from TistoryResponse to TistoryError.
_error_handler() will now return 'unknown' if the message isn't known.

The API URL creation is moved from TistoryClassCall to TistoryRequest.
It's now possible to access the uriparts attribute in the TistoryResponse
object.

Removed the TistoryAPI class and moved the API_URL template to
TistoryRequest

TistoryRequest.req was renamed to TistoryRequest.request
TistoryRequest.bytes was renamed to TistoryRequest.content to reflect
the name in requests.response

Various documentation rewrites.